disable apts daily service
you know what would be grand. if apt was a daemon. like say packagekit.
so this crap wouldn't establish an exclusive lock on the database. that'd
be really grand for not pissing me off. every single time a unit does
anything with apt it ends up pissing me off because it always locks the db
and guess what, if I want to install a package I don't give a fuck if some
twit installed a unit with their package. I am so fed up with debian
systems.