add support for enabling debug repos
on ubuntu for example the debug symbols are stripped out into a separate
repo containing only the symbols because on ubuntu itself they are
largely moot since they have remote retracing capabilities built into
their apport crash reporting system.
as such we need a way to enable this repo if necessary. to do this as
generically as possible we use an appstream repository component to
describe a repo, the component gets associated with a package same as
any other, and the package is in charge of somehow getting the repo set
up. this detaches the distro specifics of how a repo gets added as from
our pov this is a regular package install. we can do that easily via
packagekit already.

currently this only works for a single repo id but would be easy to
extend for multiple ids with some refactoring of the enabler class
(which currently only looks at one component and also asserts there is
only one)

cmake cache setting DebugRepoAppStreamIDs is the way for distros to set
the ids for installation