gather extensive diagnostics when file resolution fails
on usr-merged systems the paths randomly are /lib instead of /usr/lib,
not even consistently.

when this happens the resolution to a package name fails and
consequently every other resolver as well.
should this happen run a diagnostic script to gather up mountains of
data we can use in bug reports to figure out why the paths are wonky to
begin with

this extends the filepage with a textarea that will display the script
output (if diagnostics ran that is).

CCBUG: 438261