fix usr merge file resolution
I've finally thought of a good way to deal with this.

what happens is that sometimes we get unexpected input paths on
usr-merged systems (where /{bin,lib} are symlinked to /usr/{bin,lib})
screwing up path resolution because dpkg internally still knows the
correct path of the files (e.g. /lib/foo.so is not the same as
/usr/lib/foo.so as ar as dpkg/apt/packagekit are concerned).

to deal with this conundrum the fileresolver now explicitly executes a
usr-merge mangling attempt. i.e. if the file had /usr prefix it will be
dropped, if not it will be prefixed. this means that if /lib/foo.so
fails to resolve /usr/lib/foo.so will be tried as well. only if both
paths fail to resolve is the resolution as a whole considered failed.

should the usr-merged variant work we'll furthermore also update our
file object accordingly so the UI reflects the fixed up path...
shouldn't really have any practical impact other than being neater this
way.